Fixed uncaught errors
This version caught a previously unknown error where the program seemed
to function fine, but after physical verification, failed most of the
time. The culprit was deleting the first element in an array, when the
entire array needed to be cleaned for reuse.
Other comments were fixed, variable names were improved, and a valid
character checker was added to prevent the crashing of the program if a
character wasn’t in the Dictionary of allowed characters.